+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/datasource/UnionDataSource.java
@@ -2,9 +2,6 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.druid.model.datasource;
 
-import static com.yahoo.bard.webservice.web.ErrorMessageFormat.INVALID_DATASOURCE_UNION;
-
-import com.yahoo.bard.webservice.data.dimension.Dimension;
 import com.yahoo.bard.webservice.druid.model.query.DruidQuery;
 import com.yahoo.bard.webservice.table.ConstrainedTable;
 
@@ -15,7 +12,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * Represents a Druid Union data source.
@@ -27,33 +23,10 @@ public class UnionDataSource extends DataSource {
     /**
      * Constructor.
      *
-     * @param physicalTables  The physical tables of the data source
+     * @param physicalTable  The physical table of the data source
      */
-    public UnionDataSource(Set<ConstrainedTable> physicalTables) {
-        super(DefaultDataSourceType.UNION, physicalTables);
-        // Check whether or not our union'd tables' dimensions match
-        physicalTables.forEach(table ->
-                // For each table, cycle through its dimensions and extract its physical name (i.e. expected)
-                table.getDimensions().forEach(dimension -> {
-                    String apiName = dimension.getApiName();
-                    String expectedName = table.getPhysicalColumnName(apiName);
-                    // For every other table, cycle through their dimensions and compare their names to expected
-                    physicalTables.forEach(altTable -> {
-                        Set<String> otherNames = altTable.getDimensions().stream()
-                                .map(Dimension::getApiName)
-                                .collect(Collectors.toSet());
-                        String actualName = altTable.getPhysicalColumnName(apiName);
-                        // If this other table contains the apiName, ensure that the physical name mapping is identical
-                        if (otherNames.contains(apiName)
-                                && !expectedName.equals(actualName)) {
-                            LOG.error(INVALID_DATASOURCE_UNION.logFormat(apiName, expectedName, actualName));
-                            throw new IllegalStateException(
-                                    INVALID_DATASOURCE_UNION.format(apiName, expectedName, actualName)
-                            );
-                        }
-                    });
-                })
-        );
+    public UnionDataSource(ConstrainedTable physicalTable) {
+        super(DefaultDataSourceType.UNION, physicalTable);
     }
 
     @Override