+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/PartialDataHandlerSpec.groovy
@@ -3,13 +3,18 @@
 package com.yahoo.bard.webservice.data
 
 import static com.yahoo.bard.webservice.data.time.DefaultTimeGrain.DAY
+import static com.yahoo.bard.webservice.data.time.DefaultTimeGrain.MONTH
 import static com.yahoo.bard.webservice.data.time.DefaultTimeGrain.WEEK
+import static com.yahoo.bard.webservice.data.time.DefaultTimeGrain.YEAR
+import static com.yahoo.bard.webservice.util.IntervalUtilsSpec.buildIntervalList
+import static com.yahoo.bard.webservice.util.IntervalUtilsSpec.complementExpectedSets
 import static org.joda.time.DateTimeZone.UTC
 
 import com.yahoo.bard.webservice.data.dimension.Dimension
 import com.yahoo.bard.webservice.data.dimension.DimensionDictionary
 import com.yahoo.bard.webservice.data.dimension.impl.KeyValueStoreDimension
 import com.yahoo.bard.webservice.druid.model.query.AllGranularity
+import com.yahoo.bard.webservice.druid.model.query.Granularity
 import com.yahoo.bard.webservice.druid.model.query.GroupByQuery
 import com.yahoo.bard.webservice.metadata.TestDataSourceMetadataService
 import com.yahoo.bard.webservice.table.Column
@@ -21,7 +26,7 @@ import org.joda.time.DateTimeZone
 import org.joda.time.Interval
 
 import spock.lang.Specification
-
+import spock.lang.Unroll
 /**
  * Tests for the Partial Data Handler
  */
@@ -93,8 +98,7 @@ class PartialDataHandlerSpec extends Specification {
 
         expect:
         expectedIntervals == partialDataHandler.findMissingTimeGrainIntervals(
-                dataSourceConstraint,
-                [table] as Set,
+                table.getAvailableIntervals(dataSourceConstraint),
                 new SimplifiedIntervalList(dataSourceConstraint.getIntervals()),
                 dataSourceConstraint.getRequestGranularity()
         )
@@ -108,8 +112,7 @@ class PartialDataHandlerSpec extends Specification {
 
         expect:
         requestedIntervals == partialDataHandler.findMissingTimeGrainIntervals(
-                dataSourceConstraint,
-                [table] as Set,
+                table.getAvailableIntervals(dataSourceConstraint),
                 new SimplifiedIntervalList(dataSourceConstraint.getIntervals()),
                 dataSourceConstraint.getRequestGranularity()
         )
@@ -117,5 +120,70 @@ class PartialDataHandlerSpec extends Specification {
 
     SimplifiedIntervalList buildIntervals(List<String> intervals) {
         intervals.collect({ (new Interval(it)) }) as SimplifiedIntervalList
+    }
+
+
+    @Unroll
+    def "Complement of #supply yields #expected with fixed request and grain"() {
+        setup:
+        SimplifiedIntervalList supplyIntervals = buildIntervalList(supply)
+        SimplifiedIntervalList requestedIntervals = buildIntervalList(['2014/2015'])
+        Granularity granularity = MONTH
+        SimplifiedIntervalList expectedIntervals = buildIntervalList(expected)
+
+        expect:
+        PartialDataHandler.collectBucketedIntervalsNotInIntervalList(
+                supplyIntervals,
+                requestedIntervals,
+                granularity
+        ) == expectedIntervals
+
+        where:
+        supply               | expected
+        ["2013-04/2014-04"]  | ["2014-04/2015"]
+        ["2014-04/2014-05"]  | ["2014/2014-04", "2014-05/2015"]
+        ["2012-09/2016-05"]  | []
+        ["2012/2013"]        | ["2014/2015"]
+        ["2020/2023"]        | ["2014/2015"]
+    }
+
+
+    @Unroll
+    def "Collect of #requestedIntervals by #grain yields #expected when fixed supply is removed"() {
+        setup:
+        SimplifiedIntervalList supply = buildIntervalList(['2012-05-04/2017-02-03'])
+        SimplifiedIntervalList expectedIntervals = buildIntervalList(expected)
+        SimplifiedIntervalList requestedIntervals = buildIntervalList(requested)
+        Granularity granularity = grain
+
+        expect:
+        PartialDataHandler.collectBucketedIntervalsNotInIntervalList(
+                supply,
+                requestedIntervals,
+                granularity
+        ) == expectedIntervals
+
+        where:
+        grain | requested                 | expected
+        YEAR  | ["2012/2017"]             | ["2012/2013"]
+        MONTH | ["2012-02/2017"]          | ["2012-02/2012-06"]
+        DAY   | ["2012-02-02/2016-05"]    | ["2012-02-02/2012-05-04"]
+        DAY   | ["2012-06/2016-05"]       | []
+        YEAR  | ["2013/2018"]             | ["2017/2018"]
+        MONTH | ["2013/2017-04"]          | ["2017-02/2017-04"]
+        DAY   | ["2012-02-03/2017-04-04"] | ["2012-02-03/2012-05-04", "2017-02-03/2017-04-04"]
+    }
+
+    def "Test that complement cuts out the correct hole(s)"() {
+        given:
+        SimplifiedIntervalList from = buildIntervalList(fromAsStrings)
+        SimplifiedIntervalList remove = buildIntervalList(removeAsStrings)
+        SimplifiedIntervalList expected = buildIntervalList(expectedAsStrings)
+
+        expect:
+        PartialDataHandler.collectBucketedIntervalsNotInIntervalList(remove, from, granularity) == expected
+
+        where:
+        [granularity, comment, fromAsStrings, removeAsStrings, expectedAsStrings] << complementExpectedSets()
     }
 }