+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/handlers/PartialDataRequestHandlerSpec.groovy
@@ -8,10 +8,11 @@ import static com.yahoo.bard.webservice.web.responseprocessors.ResponseContextKe
 import com.yahoo.bard.webservice.data.PartialDataHandler
 import com.yahoo.bard.webservice.data.metric.mappers.PartialDataResultSetMapper
 import com.yahoo.bard.webservice.data.metric.mappers.ResultSetMapper
+import com.yahoo.bard.webservice.data.time.DefaultTimeGrain
 import com.yahoo.bard.webservice.druid.model.datasource.DataSource
 import com.yahoo.bard.webservice.druid.model.query.GroupByQuery
-import com.yahoo.bard.webservice.table.ConcretePhysicalTable
 import com.yahoo.bard.webservice.table.ConstrainedTable
+import com.yahoo.bard.webservice.table.PhysicalTable
 import com.yahoo.bard.webservice.table.PhysicalTableDictionary
 import com.yahoo.bard.webservice.util.SimplifiedIntervalList
 import com.yahoo.bard.webservice.web.DataApiRequest
@@ -23,8 +24,6 @@ import org.joda.time.Interval
 
 import spock.lang.Specification
 
-import java.util.stream.Stream
-
 import javax.ws.rs.core.MultivaluedHashMap
 import javax.ws.rs.core.MultivaluedMap
 
@@ -34,7 +33,7 @@ class PartialDataRequestHandlerSpec extends Specification {
 
     DataRequestHandler next = Mock(DataRequestHandler)
     PhysicalTableDictionary physicalTableDictionary = Mock(PhysicalTableDictionary)
-    Set<ConcretePhysicalTable> physicalTables = [Mock(ConstrainedTable)] as Set
+    PhysicalTable physicalTable = Mock(ConstrainedTable)
     PartialDataHandler partialDataHandler = Mock(PartialDataHandler)
 
     RequestContext rc = Mock(RequestContext)
@@ -42,6 +41,8 @@ class PartialDataRequestHandlerSpec extends Specification {
     GroupByQuery groupByQuery = Mock(GroupByQuery)
     DataSource dataSource = Mock(DataSource)
     MappingResponseProcessor response = Mock(MappingResponseProcessor)
+    SimplifiedIntervalList availableIntervals
+
 
     PartialDataRequestHandler handler = new PartialDataRequestHandler(
             next,
@@ -49,16 +50,17 @@ class PartialDataRequestHandlerSpec extends Specification {
     )
 
     def setup() {
+        availableIntervals = new SimplifiedIntervalList([new Interval(0, 15), new Interval(30, 1000)])
+        physicalTable.getAvailableIntervals() >> availableIntervals
         apiRequest.getDimensions() >> Collections.emptySet()
         apiRequest.getFilterDimensions() >> Collections.emptySet()
         apiRequest.getFilters() >> Collections.emptyMap()
+        apiRequest.getGranularity() >> DefaultTimeGrain.DAY
         groupByQuery.getMetricDimensions() >> Collections.emptySet()
         groupByQuery.getDependentFieldNames() >> Collections.emptySet()
         groupByQuery.getInnermostQuery() >> groupByQuery
         groupByQuery.getDataSource() >> dataSource
-        //physicalTableDictionary.get(_) >> physicalTables[0]
-        //dataSource.getNames() >> ["name"]
-        dataSource.getPhysicalTables() >> physicalTables
+        dataSource.getPhysicalTable() >> physicalTable
     }
 
     def cleanup() {
@@ -78,7 +80,7 @@ class PartialDataRequestHandlerSpec extends Specification {
         then:
         success
         1 * partialDataHandler.findMissingTimeGrainIntervals(
-                _ as Stream<ConstrainedTable>,
+                availableIntervals,
                 new SimplifiedIntervalList(apiRequest.intervals),
                 apiRequest.granularity
         ) >> intervals
@@ -109,7 +111,7 @@ class PartialDataRequestHandlerSpec extends Specification {
         then:
         success
         1 * partialDataHandler.findMissingTimeGrainIntervals(
-                _ as Stream<ConstrainedTable>,
+                availableIntervals,
                 new SimplifiedIntervalList(apiRequest.intervals),
                 apiRequest.granularity
         ) >> nonEmptyIntervals
@@ -153,7 +155,7 @@ class PartialDataRequestHandlerSpec extends Specification {
         then:
         success
         1 * partialDataHandler.findMissingTimeGrainIntervals(
-                _ as Stream<ConstrainedTable>,
+                availableIntervals,
                 new SimplifiedIntervalList(apiRequest.intervals),
                 apiRequest.granularity
         ) >> nonEmptyIntervals